% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhePhebAlzhiemerCohortDiagnostics.R
\docType{package}
\name{PhePhebAlzhiemerCohortDiagnostics-package}
\alias{PhePhebAlzhiemerCohortDiagnostics}
\alias{PhePhebAlzhiemerCohortDiagnostics-package}
\title{PhePhebAlzhiemerCohortDiagnostics: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
